Server: Normalize rotational vector
If we were to move on with having
rotation data and expected it to be in range of -1-1
then we could get into trouble and have exploit
on our hands if hacked client sent vector out of this
range.
Thus we have to normalize the vector as it comes in.
I have tried the Fast Inverse Square Root algo
and while it does work it can produce slight deviation which is bad
in the case of aiming.