Server: Add grenade floating blocks destroying
This is basically as fast as we can do it.
It doesnt look nice to the eye but manual assigment
of the array had to be done because of the unusual
positioning.

I added also a check at the start of the checkNode
function to check for empty block.
While we could have have just looped trough 3D for loop
and grab neighbours thats around 729 we would have to preform.
This method reduces us to only 54.