Source: Merge PacketReceive and Packet files
This merges PacketReceive.c/h and Packets.c/h
and while doing so removes writeBlockLine from
Map.c and removes ColorConversion.h

Removal of writeBlockLine is due to the fact that this is mostly
protocol thing and also it was used only once for 1 thing.
May become static inline function later for readability sake.

The same applies for ColorConversion.h removal.
Contained only 1 function that was used in 1 file after the merge.

To be honest i hated having packets handling split into receive and send files.
So this does make me quite happy :)