Source: Redo the login.
With this there is a possibility to have as many
roles as one could wish tho i still think the 5 default
ones are more than enough for basically anyone really.

There is only 1 change for end users with this change.
That is that moderators can login only using mod as role.