Source: Fix loading with names with space in ban list
When a name had a space inside and was placed into ban list it would cause all the
players loading into game after this point to not finish loading
and thus never get into game.

Fix this by loading everything until we hit , or are limited by char limit of name.