Protocol.c: Check even block between corners
There is an edge condition where if the grenade ends up at Z
level of 1 and would destroy part of platform then the previous
if conditions would still allow this as they were checking only
corner blocks.

This is no longer the case with this commit as we now
fully check all destroyable corner Z levels of thrown grenade.