Protocol.c: break for loop when we find the visited node in the list
This is around 2x fast now on bigger.

WTF was i thinking even writting this cursed thing ?
Sure may have been 2am code that did it. But for real.

While this commit speeds it up a lot. It doesnt even come close
to what it will be once this is replaced.