Packets.c: Ban ZeroSpades
This has not been an easy decision to do.
But creator of ZeroSpades has been asked multiple
times to upload the code of his client to github
or share it publically (Or with people that got the binary)
in some sort of way.

Creator is willing to share this source code
upon completion of the client in some sense
but even now there are compiled versions of
the client being used by people.
This is still a violation of the GPLv3 license.

And thus we have decided to temporarily ban
the client until the source code is released
to the public.
And creator promised that they will release it
once they finish it. So lets hope the ban is
a short one :)