Packets.c: Do not allow shooting or placing blocks when sprinting
When sprinting the player is not allowed to
either shoot, place blocks, throw grenades
and thus lets not allow such actions to occur.

There could be a really rare chance that hit
or place packet could arrive sooner then
keyboard input packet that cancels sprinting
but this is rather unlikely to happen.