Packets.c: Check for other timers in block action
Before we were only checking single timer for each
block action type. But that is not the case in real
life. We cant have player destroy 3 blocks and then
instantly after that destroy a single block with no
delay. That doesnt make sense.

There is a delay with animations for non modified
clients.

While at it move the delays into enum so they are
easier to understand and change them to more correct values
(Have to be checked if completely correct)