Types.h: Make LOG_X compatible with Clang
The previously used ##__VA_ARGS__ is used
in macros when normal __VA_ARGS__ cant be.
Such as removing remaining , if arguments
are empty.

But ## is a GNU extension and Clang would
complain about that fact with a warning.
And since we have warnings as errors.
Clang would create an error.

Thus we have to make variadic macros compatible
with C99 standart and not use ##.

This cant be done in a simple manner and thus an
internal (INT) LOG functions have to be used.
One for printf and second for fprintf.
This way we can workaround the limitation
and hope that __VA_OPT__ becomes available
in the upcoming C standart release which would
resolve this hacky workaround.