Server: Get player unstuck
If a player is baricaded by other player
into blocks usually they would have to dig out these blocks
but that may not be fast enough.

If a player gets stuck like this (May be noclip)
try to find an block in 3x3 range where the player could
crouch if they were crouching or standing if they were
standing and set their position there if such a block is found.

If none is found it would set their position
to some block in team spawn.