Server: Make grenades into linked list
This logic allows us to have as many grenades per player
as they wish.

Also do NOT create packets when the number of players
on server is 0. Its pointless. We have nobody to send it
to causing pointless malloc and then freeing of the memory
if it were not sent. This was found during testing to make
sure grenades explode even when player disconnects while grenade
is in the air and nobody is on the server.
Sending the grenade packet would allocate memory for the packet but
it would not be sent to anybody causing a memory leak since
it didnt get freed as it does after sending it.