Source: Add server console
Server console is a console
that server administrator can use to use
commands without being actually connected
in game.

For now the console includes also
history of commands preformed during
that session of the server.
Meaning when you stop the server
history will be cleaned.

Tab completion will be implemented
later on.