Packets.c: Normalize vector locally to player
This has few advantages over the old method. 1
being that it doesnt need to normalize vectors
that were slightly over 1 and thus screwing up
throwing of grenades in rare scenarios (Jumping and up orientation)
But also makes sure that if player has speedhacks grenades
wont be affected by it as much as they would be.