Source: Add block buffer for map loading players
In the rare instance that player cant load the map
super fast or player breaks a block in the time
the other player is loading the map these changes would
not end up on that clients map which results in mismatch.

While unlikely it can happen and thus should be handled.
This saves all block actions and block lines that would get
sent to the player but cant be and are then replayer with
proper color once map finishes loading.