Packets: SetTool.c: Complete the reload cycle when set tool arrives
Normally we would just not send the last packet and clients would
not finish reloading. This isnt the case in BetterSpades and we
have to force the client to finish reloading and also set the right
number of bullets in clip and reserve.

We do this by sending the final packet including the number of
bullets we had before reloading finished. Since we didnt finish
reloading fully the clip and reserve has not been modified client
side and we just send an ordinary packet thus not breaking other
clients and avoiding potential issues from this BetterSpades bug.