Source: Remove global compressed map
Until now we had a weird system in place where
we would compress the map globally and then
once a player joined compress it again regardless.
Also it was possible to have some weird memory behaviour
to occur which we ofcourse do not want.

This makes the map PER PLAYER. So each player compresses the map.
The main advantage being that there is 0 collisions between map loading
for players.