Notice.c: Add UTF-8 magic byte for OpenSpades
OpenSpades requires the server to send messages with 0xFF at the beginning
in order for them to be treated as UTF-8, otherwise they are treated as CP437.
Not sending the byte causes unicode text to show up as garbled. Unfortunately
BetterSpades and VOXLAP don't support UTF-8 yet, so this only sends the magic
byte for OpenSpades.