Server: Check for invalid floats in all packets
This adds checks for nan/inf in all instances where a float is read
from a packet, and drops it if an invalid value is found.

The vector3f_t check is located in a new file, VectorChecks.c.

Credits to burner (DavidCo113 <davidco7777@protonmail.com>) for telling me about this.