PositionChecks.c: Rewrite valid_player_pos
This does allow for wider range of things then the previous
function but at the same time its faster and FAR easier to
read.

Previous implementation followed the principle that values below 0
were not allowed unless specific situation occured.
While this was great it did leave a lot of room for error and thus
we decided to allow the full range of motion for the player
without the need for specific situation (Other then water
interaction and crouching)