Server: Use sigaction to handle signals
This commit restructures how signal handling is done in SpadesX, e.g. handling SIGINT.

This commit addresses two issues: Firstly, SIGTERM is now handled, which is particularly useful for Docker. Secondly, the deprecated use of the signal function has been replaced with sigaction.

To handle other signals, all that's required is to expand the switch statement - clean and only a few lines of code.