ExistingPlayer.c: Convert `strcpy` -> `strncpy`; add null-terminated safety
`strcpy` is unsafe and has thus been replaced with `strncpy`, which will
at least accept a maximum length parameter. `strncpy` may have other
flaws, such as the lack of null termination, which will be discussed in
a moment. Thus, one may want to change the usage of `strncpy` here in
the future if there is a better function that doesn't require a
dependency like `libbsd` as we would rather do without.

`strcpy` and `strncpy` do not add null termination to the end of their
copied strings, so this has been added as a safety precaution as per
recommendations by @rndtrash and @Haxk20.