Rewrite 'unique username system' (#86)
* ExistingPlayer.c: Rewrite unique username system

The code used to ensure usernames are unique for online players has been
largely rewritten in this commit.

The major takeaways are:

- This commit aims to resolve #72; it appears to have no trouble modifying
  players' usernames to be unique regardless of the length of their
  desired name or their player ID as a string.

- Removed use case of `libbsd`, as this new version does not use `strlcat`.
    - Some other cleanup may be needed with CMake and documentation to
      finish removing this now-unused dependency.

- I'm almost certain there is a lot to improve with this code - C
  strings are my greatest weakness, lol. Please lend a hand in updating
  the code as you see fit.
    - I'm particularly worried about the use of standard string
      functions that aren't aware of a max bound they should restrict
      themselves to. Though, if they're not needed, then they're not
      needed - I'll leave this to the professionals to decide.

Thanks

* Server: Define player name string length constants

* ExistingPlayer.c: Convert `strcpy` -> `strncpy`; add null-terminated safety

`strcpy` is unsafe and has thus been replaced with `strncpy`, which will
at least accept a maximum length parameter. `strncpy` may have other
flaws, such as the lack of null termination, which will be discussed in
a moment. Thus, one may want to change the usage of `strncpy` here in
the future if there is a better function that doesn't require a
dependency like `libbsd` as we would rather do without.

`strcpy` and `strncpy` do not add null termination to the end of their
copied strings, so this has been added as a safety precaution as per
recommendations by @rndtrash and @Haxk20.

* ExistingPlayer.c: Use safer unsigned specifier for name length

* Server: Remove duplicate length constant; add usage in

* All: Remove libbsd dependency

Compiles and runs on my machine.

* ExistingPlayer.c: Use `strncmp`, remove redundant string operation

We can use `strncmp` in the modified places of this commit as the player's name length is known and may as well be used for additional safety.

`new_name[0] = '\0'` shouldn't be necessary on line 133 due to line 113 already initialising the string with a guaranteed null character inside it.